@@ -5,21 +5,22 @@ author: Denis Bakhtin <denis.bakhtin@gmail.com>
 homepage: https://github.com/denisbakhtin/dirk
 
 environment:
-  sdk: ">=2.10.0 <3.0.0" 
+  sdk: ">=2.12.0 <3.0.0" 
 
 dependencies:
-  analyzer: ^0.41.2
-  build: ^1.6.2
+  analyzer: ^1.2.0
+  build: ^2.0.0
   build_config: ^0.4.6
   string_scanner: ^1.1.0
   build_runner: ^1.10.6
   meta: ^1.2.4
-  dart_style: ^1.3.10
+  dart_style: ^1.3.1
   path: ^1.7.0
   universal_io: ^1.0.1
   recase: ^3.0.1
   charcode: ^1.2.0
   glob: ^2.0.0
+  source_gen: ^0.9.10+4
   
 dev_dependencies:
   test: ^1.15.7
