@@ -1,9 +1,13 @@
-//import 'package:todo_reporter_generator/annotations.dart';
-import 'views/code/views.dirk.dart';
+//import 'views/code/views.dirk.dart';
+import 'package:dirk/annotations.dart';
 
+part 'todo.g.dart';
+
+@DirkViews("lib/views/")
 class Todo {
   @override
   String toString() {
-    return IndexView();
+    return "";
+    //return "IndexView()";
   }
 }
