@@ -29,7 +29,7 @@ class DirkAST {
   DirkAST({
     String start = '',
     String layout = '@renderBody()',
-    String contents,
+    String contents = '',
     this.isPartial = false,
   })  : this.start = start,
         this.contents = isPartial
@@ -126,9 +126,9 @@ class DirkAST {
         //matches: if (...) {...}
         if (_scanner.scan(_if)) {
           if (_scanner.scan(_ifCondition)) {
-            var condition = _scanner.lastMatch.group(1);
+            var condition = _scanner.lastMatch?.group(1);
             var body = _scanTillClosingCurlyBrace(_scanner);
-            tree.add(Token(TokenType.$if, condition));
+            tree.add(Token(TokenType.$if, condition ?? ""));
             _deepParse(tree.last.children, body);
           } else {
             errors.add(DirkError(
@@ -139,9 +139,9 @@ class DirkAST {
         //matches: for(...) {...}
         if (_scanner.scan(_for)) {
           if (_scanner.scan(_forCondition)) {
-            var condition = _scanner.lastMatch.group(1);
+            var condition = _scanner.lastMatch?.group(1);
             var body = _scanTillClosingCurlyBrace(_scanner);
-            tree.add(Token(TokenType.$for, condition));
+            tree.add(Token(TokenType.$for, condition ?? ""));
             _deepParse(tree.last.children, body);
           } else {
             errors.add(DirkError(
@@ -174,12 +174,12 @@ class DirkAST {
         }
         //matches: anyExpression
         if (_scanner.scan(_expr)) {
-          var body = _scanner.lastMatch.group(1);
-          tree.add(Token(TokenType.expression, body));
+          var body = _scanner.lastMatch?.group(1);
+          tree.add(Token(TokenType.expression, body ?? ""));
           while (!_scanner.isDone) {
             //.member match
             if (_scanner.scan(_member)) {
-              tree.last.content += _scanner.lastMatch.group(1);
+              tree.last.content += _scanner.lastMatch?.group(1) ?? "";
               continue;
             }
             //( match
@@ -224,11 +224,11 @@ class DirkAST {
     StringBuffer result = StringBuffer();
     List<Token> imports =
         tree.where((el) => el.type == TokenType.$import).toList();
-    Token model =
-        tree.firstWhere((el) => el.type == TokenType.model, orElse: () => null);
+    Token model = tree.firstWhere((el) => el.type == TokenType.model,
+        orElse: () => Token(TokenType.model, ""));
     tree.removeWhere(
         (el) => el.type == TokenType.$import || el.type == TokenType.model);
-    String modelType = model?.content ?? "";
+    String modelType = model.content;
 
     result.writeln("// Generated by DirkAST, don't modify");
     result.writeln("");
